requirements.txt
weasyprint
django-crontab
pillow